requirements-stable.txt
flask>=2.3.0,<3.0.0
flask-cors>=4.0.0,<5.0.0
requests>=2.31.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
fpdf2>=2.8.0,<3.0.0
matplotlib>=3.7.0,<4.0.0
pillow>=10.0.0,<11.0.0
numpy>=1.24.0,<2.0.0
python-telegram-bot>=13.15,<14.0.0
openai>=1.3.0,<2.0.0
supabase>=2.0.0,<3.0.0
kaleido>=0.2.0,<1.0.0
python-dateutil>=2.8.0,<3.0.0
